macaca-report 1.0-SNAPSHOT version

+++ src/main/java/com/macacajs/reporter/models/StatsDataModel.java
@@ -36,7 +36,7 @@ public class StatsDataModel {
     /**
      * 通过占比
      */
-    private Long passPercent;
+    private double passPercent;
     /**
      * 挂起占比
      */
